Extract shared class-var-field-assignment rejection helper
generate_field_assignment_open's and the new conditional-branch C1
arm's class-var-write rejection were duplicated verbatim (flagged by
review on PR #3359) rather than sharing an implementation, violating
CLAUDE.md's no-duplicate-implementations rule. Extracts
reject_class_var_field_assignment (util.rs) and calls it from both
sites.
